requirements.txt
google-genai
playwright
termcolor
